requirements.txt
cffi~=1.16.0
Pillow~=10.1.0
pip~=21.3.1
wheel~=0.37.1
rsa~=4.9
pyasn1~=0.5.0
setuptools~=60.2.0
gast~=0.5.4
numpy~=1.26.1
h5py~=3.10.0
pytz~=2023.3.post1
keras~=2.14.0
tensorflow~=2.14.0
six~=1.16.0
scipy~=1.11.3
wrapt~=1.14.1
requests~=2.31.0
numba~=0.58.1
llvmlite~=0.41.1
pooch~=1.8.0
packaging~=23.2
platformdirs~=3.11.0
matplotlib~=3.8.1
Django~=4.2.7
asgiref~=3.7.2
sqlparse~=0.4.4
cachetools~=5.3.2
certifi~=2023.7.22
urllib3~=2.0.7
joblib~=1.3.2
soundfile~=0.12.1
audioread~=3.0.1
soxr~=0.3.7
msgpack~=1.0.7
decorator~=5.1.1
scikit-learn~=1.3.2
threadpoolctl~=3.2.0
python-dateutil~=2.8.2
Markdown~=3.5.1
oauthlib~=3.2.2
idna~=3.4
watchdog~=2.1.9
MarkupSafe~=2.1.3
werkzeug~=3.0.1
contourpy~=1.2.0
fonttools~=4.44.0
pycparser~=2.21
pyparsing~=3.1.1
termcolor~=2.3.0
librosa~=0.10.1
pydub~=0.25.1
birdnetlib~=0.12.0
cycler~=0.12.1
kiwisolver~=1.4.5
tensorboard~=2.14.1
flatbuffers~=23.5.26
astunparse~=1.6.3
djangorestframework~=3.14.0
resampy~=0.4.2